pyproject.toml
[project]
name = "dspy-examples"
version = "0.1.0"
description = "DSPy prompt optimization examples"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dspy-ai>=2.5.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
markers = [
    "integration: marks tests as requiring external services (deselect with '-m \"not integration\"')",
]

[tool.hatch.build.targets.wheel]
packages = ["src/dspy_examples"]